refactor: extract GuidanceEventRouter to bring plugin class below 1,000 LOC
Moves 9 @Subscribe handlers (onActorDeath, onAnimationChanged, onNpcSpawned,
onNpcDespawned, onInteractingChanged, onHitsplatApplied, onWidgetLoaded,
onMenuEntryAdded, onMenuOptionClicked) plus two dead private methods
(authoringLog, logContainerChange, which were superseded by AuthoringLogger)
into the new GuidanceEventRouter singleton.

Plugin drops from 1,281 to 904 LOC. Router is 452 LOC.
Part of A2 in docs/ROADMAP.md (Tier A).
